+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Notable changes to this project will be documented in this file.
 ## [2.0.0]
 - Support for SilverStripe 4
 - Rewrite with new Asset backend
+- Extend Upload rather than Image
+- Add option to scale only height or width
 
 ## [1.0.1]